requirements.txt
chevron>=0.14.0
pillow>=10.3.0
